Add ability to update story from scum master screen and refactor
