+++ ScrumStorySizer.Library/Components/StandardStory.razor
@@ -1,28 +1,6 @@
-@inherits StoryBase
-
 <div class="row">
     <div class="col">
         <label style="width:100%">Story Name</label>
         <input autofocus class="form-control" style="max-width:600px;" type="text" value="@_storyName" @oninput="UpdateStoryName" @onfocusin="() => _titleFocus = true" @onfocusout="() => { _titleFocus = false; ReconcileStoryName(); }" />
-    </div>
-</div>
-<div class="mt-4 d-flex">
-    <div class="btn-group" role="group" style="width:680px;" aria-label="Admin group">
-        <button type="button" class="btn btn-primary btn-admin" disabled="@ShowResultsDisabled"
-            @onclick="RevealVotes">Show Results</button>
-        @if (PokerVote.TimeLeft > 0)
-        {
-            <button type="button" class="btn btn-primary btn-admin" @onclick="CancelTimer">Cancel Timer
-                (@PokerVote.TimeLeft)</button>
-        }
-        else
-        {
-            <button type="button" title="Start 10 second timer." class="btn btn-primary btn-admin"
-            disabled="@ShowResultsDisabled" @onclick="() => StartTimer(10)">Start Timer</button>
-        }
-        <button type="button" class="btn btn-primary btn-admin" @onclick="ClearVotes">Clear
-            Votes</button>
-        <button type="button" class="btn btn-primary btn-admin" @onclick="NewStory">New
-            Story</button>
     </div>
 </div>
