+++ ScrumStorySizer.Library/Pages/Master.razor
@@ -6,14 +6,85 @@
         <hr />
         @if (DevOpsCredential?.IsEnabled == true)
         {
-            <DevOpsStory DevOpsCredential="DevOpsCredential" />
+            <DevOpsStory DevOpsCredential="DevOpsCredential" @ref="_devOpsStoryRef" />
         }
         else
         {
             <StandardStory />
         }
+        <div class="mt-4 d-flex">
+            <div class="btn-group" role="group" style="width:680px;" aria-label="Admin group">
+                <button type="button" class="btn btn-primary btn-admin" disabled="@ShowResultsDisabled"
+                    @onclick="RevealVotes">Show Results</button>
+                @if (PokerVote.TimeLeft > 0)
+                {
+                    <button type="button" class="btn btn-primary btn-admin" @onclick="CancelTimer">Cancel Timer
+                        (@PokerVote.TimeLeft)</button>
+                }
+                else
+                {
+                    <button type="button" title="Start 10 second timer." class="btn btn-primary btn-admin"
+                    disabled="@ShowResultsDisabled" @onclick="() => StartTimer(10)">Start Timer</button>
+                }
+                <button type="button" class="btn btn-primary btn-admin" @onclick="ClearVotes">Clear
+                    Votes</button>
+                <button type="button" class="btn btn-primary btn-admin" @onclick="NewStory">New
+                    Story</button>
+            </div>
+        </div>
     </div>
 </div>
+
+@if (DevOpsCredential?.IsEnabled == true)
+{
+    <div class="card bg-dark">
+        <div class="card-body">
+            <h5 class="card-title">Story Details</h5>
+            <hr />
+            <div class="row">
+                <div class="col">
+                    <label>Story Name</label>
+                    <h4>@(string.IsNullOrWhiteSpace(PokerVote.WorkItem?.Title) ? "Waiting..." : PokerVote.WorkItem.Title)
+                    </h4>
+                </div>
+            </div>
+            @if (PokerVote.ShowVotes && PokerVote.WorkItem?.Id != null)
+            {
+                <div title="Picking a size will automatically save the PBI.">
+                    <div class="d-flex">
+                        <label class="mx-auto">Set Story Size</label>
+                    </div>
+                    <div class="vote-grp-lg d-flex mx-auto">
+                        <div class="btn-group mx-auto" role="group" style="width:500px;" aria-label="Voting group">
+                            <button type="button" title="Story Size 1" class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.XS)">XS</button>
+                            <button type="button" title="Story Size 2" class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.S)">S</button>
+                            <button type="button" title="Story Size 3" class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.M)">M</button>
+                            <button type="button" title="Story Size 5" class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.L)">L</button>
+                            <button type="button" title="Story Size 8" class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.XL)">XL</button>
+                            <button type="button" title="Story Size 13. Needs to be broken up." class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.XXL)">XXL</button>
+                        </div>
+                    </div>
+                    <div class="vote-grp-sm">
+                        <div class="d-flex mt-4 mx-auto">
+                            <div class="btn-group mx-auto" role="group" style="width:500px;" aria-label="Voting group">
+                                <button type="button" title="Story Size 1" class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.XS)">XS</button>
+                                <button type="button" title="Story Size 2" class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.S)">S</button>
+                                <button type="button" title="Story Size 3" class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.M)">M</button>
+                            </div>
+                        </div>
+                        <div class="d-flex mt-4 mx-auto">
+                            <div class="btn-group mx-auto" role="group" style="width:500px;" aria-label="Voting group">
+                                <button type="button" title="Story Size 5" class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.L)">L</button>
+                                <button type="button" title="Story Size 8" class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.XL)">XL</button>
+                                <button type="button" title="Story Size 13. Needs to be broken up." class="btn btn-primary btn-vote" @onclick="() => SaveVote(StorySize.XXL)">XXL</button>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            }
+        </div>
+    </div>
+}
 
 <div class="card bg-dark">
     <div class="card-body">
